fix: 服务价格支持小数与批量价格可选参数

+++ admin/internal/logic/mcp/serviceUpdatePriceLogic.go
@@ -47,7 +47,7 @@ func (l *ServiceUpdatePriceLogic) ServiceUpdatePrice(req *types.ServiceUpdatePri
 	}
 
 	oldPrice := mcpService.Price
-	mcpService.Price = float64(req.Price)
+	mcpService.Price = req.Price
 	mcpService.UpdateTime = time.Now()
 
 	err = l.svcCtx.McpServiceModel.Update(l.ctx, mcpService)
